refactor: Generalize stats config handling
