+++ lib/compile.js
@@ -3,16 +3,8 @@
 const _ = require('lodash');
 const BbPromise = require('bluebird');
 const webpack = require('webpack');
-const tty = require('tty');
 const isBuiltinModule = require('is-builtin-module');
-
-const defaultStatsConfig = {
-  colors: tty.isatty(process.stdout.fd),
-  hash: false,
-  version: false,
-  chunks: false,
-  children: false
-};
+const logStats = require('./logStats');
 
 function ensureArray(obj) {
   return _.isArray(obj) ? obj : [obj];
@@ -20,10 +12,7 @@ function ensureArray(obj) {
 
 function getStatsLogger(statsConfig, consoleLog) {
   return stats => {
-    const statsOutput = stats.toString(statsConfig || defaultStatsConfig);
-    if (statsOutput) {
-      consoleLog(statsOutput);
-    }
+    logStats(stats, statsConfig, consoleLog);
   };
 }
