+++ lib/wpwatch.js
@@ -3,7 +3,7 @@
 const _ = require('lodash');
 const BbPromise = require('bluebird');
 const webpack = require('webpack');
-const tty = require('tty');
+const logStats = require('./logStats');
 
 module.exports = {
   wpwatch() {
@@ -60,14 +60,7 @@ module.exports = {
       });
     }
 
-    const consoleStats = this.webpackConfig.stats || {
-      colors: tty.isatty(process.stdout.fd),
-      hash: false,
-      version: false,
-      chunks: false,
-      children: false
-    };
-
+    const consoleStats = this.webpackConfig.stats;
     // This starts the watch and waits for the immediate compile that follows to end or fail.
     let lastHash = null;
 
@@ -98,10 +91,7 @@ module.exports = {
 
         if (stats) {
           lastHash = stats.hash;
-          const statsOutput = stats.toString(consoleStats);
-          if (statsOutput) {
-            this.serverless.cli.consoleLog(stats.toString(consoleStats));
-          }
+          logStats(stats, consoleStats, this.serverless.cli.consoleLog);
         }
 
         if (firstRun) {
